+++ lib/td_dq/implementations/bulk_load.ex
@@ -3,6 +3,8 @@ defmodule TdDq.Implementations.BulkLoad do
   Bulk Load Implementations
   """
 
+  import Ecto.Query
+
   alias Ecto.Changeset
   alias TdCache.DomainCache
   alias TdCache.TemplateCache
@@ -36,53 +38,97 @@ defmodule TdDq.Implementations.BulkLoad do
   def required_headers, do: @required_headers
 
   def bulk_load(implementations, claims) do
+    bulk_load(implementations, claims, false)
+  end
+
+  def bulk_load(implementations, claims, auto_publish) do
     Logger.info("Loading Implementations...")
 
     Timer.time(
-      fn -> do_bulk_load(implementations, claims) end,
+      fn -> do_bulk_load(implementations, claims, auto_publish) end,
       fn millis, _ -> Logger.info("Implementation loaded in #{millis}ms") end
     )
   end
 
-  defp do_bulk_load(implementations, claims) do
-    %{ids: ids} = result = create_implementations(implementations, claims)
+  defp do_bulk_load(implementations, claims, auto_publish) do
+    %{ids: ids} = result = upsert_implementations(implementations, claims, auto_publish)
     @index_worker.reindex_implementations(ids)
     {:ok, result}
   end
 
-  defp create_implementations(implementations, claims) do
+  defp upsert_implementations(implementations_params, claims, auto_publish) do
+    new_status = if auto_publish, do: "published", else: "draft"
+
+    params_by_impl_key = Enum.reduce(
+      implementations_params,
+      %{},
+      fn %{"implementation_key" => implementation_key} = ip, acc ->
+        Map.put(acc, implementation_key, Map.put(ip, "status", new_status))
+      end
+    )
+    implementation_keys = Map.keys(params_by_impl_key)
+    %{existing_implementations_struct: existing_implementations_struct, update_results: results_so_far} =
+    Implementations.Implementation
+    |> where([i], i.implementation_key in ^implementation_keys)
+    |> TdDd.Repo.all()
+    |> update_implementations(params_by_impl_key, claims)
+
+    Enum.map(existing_implementations_struct, & &1.implementation_key)
+    |> then(&Enum.filter(implementations_params, fn %{"implementation_key" => implementation_key} -> implementation_key not in &1 end))
+    |> create_implementations(results_so_far, new_status, claims)
+    |> Map.update!(:ids, &Enum.reverse/1)
+    |> Map.update!(:errors, &Enum.reverse/1)
+  end
+
+  defp make_implementations_errors(result, acc) do
+    case result do
+      {:ok, %{implementation: %{id: id}}} ->
+        %{acc | ids: [id | acc.ids]}
+
+      {:error, _, changeset, _} ->
+        error = Changeset.traverse_errors(changeset, &ErrorHelpers.translate_error/1)
+        implementation_key = Changeset.get_field(changeset, :implementation_key)
+
+        %{
+          acc
+          | errors: [%{implementation_key: implementation_key, message: error} | acc.errors]
+        }
+
+      {:error, {implementation_key, error}} ->
+        %{
+          acc
+          | errors: [
+              %{implementation_key: implementation_key, message: %{implementation: [error]}}
+              | acc.errors
+            ]
+        }
+    end
+  end
+
+  defp update_implementations(existing_implementations_struct, params_by_impl_key, claims) do
+    results = Enum.reduce(
+      existing_implementations_struct,
+      %{ids: [], errors: []},
+      fn %{implementation_key: impl_key} = existing_implementation, acc ->
+        existing_implementation
+        |> Implementations.update_implementation(params_by_impl_key[impl_key], claims)
+        |> make_implementations_errors(acc)
+      end
+    )
+    %{update_results: results, existing_implementations_struct: existing_implementations_struct}
+  end
+
+  defp create_implementations(implementations, results_so_far, new_status, claims) do
     implementations
-    |> Enum.reduce(%{ids: [], errors: []}, fn imp, acc ->
+    |> Enum.reduce(results_so_far, fn imp, acc ->
       with {:ok, imp} <- enrich_implementation(imp),
            {:ok, imp} <- maybe_put_domain_id(imp),
            {:ok, imp} <- format_df_content(imp) do
-        imp =
-          imp
-          |> Map.put("status", "draft")
-          |> Map.put("version", 1)
-
-        case create_implementation(imp, claims) do
-          {:ok, %{implementation: %{id: id}}} ->
-            %{acc | ids: [id | acc.ids]}
-
-          {:error, _, changeset, _} ->
-            error = Changeset.traverse_errors(changeset, &ErrorHelpers.translate_error/1)
-            implementation_key = Changeset.get_field(changeset, :implementation_key)
-
-            %{
-              acc
-              | errors: [%{implementation_key: implementation_key, message: error} | acc.errors]
-            }
-
-          {:error, {implementation_key, error}} ->
-            %{
-              acc
-              | errors: [
-                  %{implementation_key: implementation_key, message: %{implementation: [error]}}
-                  | acc.errors
-                ]
-            }
-        end
+        imp
+        |> Map.put("status", new_status)
+        |> Map.put("version", 1)
+        |> create_implementation(claims)
+        |> make_implementations_errors(acc)
       else
         {:error, error} ->
           %{
@@ -93,8 +139,6 @@ defmodule TdDq.Implementations.BulkLoad do
           }
       end
     end)
-    |> Map.update!(:ids, &Enum.reverse/1)
-    |> Map.update!(:errors, &Enum.reverse/1)
   end
 
   defp create_implementation(%{"rule_name" => rule_name} = imp, claims)
@@ -150,7 +194,7 @@ defmodule TdDq.Implementations.BulkLoad do
 
   defp get_domain_id_by_external_id(external_id) do
     case DomainCache.external_id_to_id(external_id) do
-      :error -> {:error, "Domain with external id #{external_id} doesn't exists"}
+      :error -> {:error, "Domain with external id #{external_id} doesn't exist"}
       domain_id -> domain_id
     end
   end
@@ -159,7 +203,7 @@ defmodule TdDq.Implementations.BulkLoad do
        when is_binary(template_name) do
     case TemplateCache.get_by_name!(template_name) do
       nil ->
-        {:error, %{"template" => ["Template #{template_name} doesn't exists"]}}
+        {:error, %{"template" => ["Template #{template_name} doesn't exist"]}}
 
       template ->
         template