+++ lib/td_dq/implementations/policy.ex
@@ -9,6 +9,7 @@ defmodule TdDq.Implementations.Policy do
 
   @workflow_actions [:delete, :edit, :move, :execute, :publish, :restore, :reject, :submit]
   @media_actions [
+    :auto_publish,
     "execute",
     "create",
     "createBasic",
@@ -53,6 +54,10 @@ defmodule TdDq.Implementations.Policy do
   def authorize("execute", %{} = claims, _params),
     do: Permissions.authorized?(claims, :execute_quality_rule_implementations)
 
+  def authorize(:auto_publish, %{} = claims, _params) do
+    Permissions.authorized?(claims, :publish_implementation)
+  end
+
   def authorize("create", %{} = claims, _params),
     do: Permissions.authorized?(claims, :manage_quality_rule_implementations)
 
@@ -93,11 +98,11 @@ defmodule TdDq.Implementations.Policy do
     do: Permissions.authorized?(claims, :manage_rule_results)
 
   def authorize(action, %{role: "admin"}, %Changeset{})
-      when action in [:create, :delete, :update, :move, :clone],
+      when action in [:create, :delete, :update, :move, :clone, :publish],
       do: true
 
   def authorize(action, %{} = claims, %Changeset{} = changeset)
-      when action in [:create, :delete, :update] do
+      when action in [:create, :delete, :update, :publish] do
     domain_id = Changeset.fetch_field!(changeset, :domain_id)
     Enum.all?(permissions(changeset), &Permissions.authorized?(claims, &1, domain_id))
   end
@@ -169,7 +174,7 @@ defmodule TdDq.Implementations.Policy do
 
   def authorize(action, %{} = claims, %Implementation{domain_id: domain_id} = implementation)
       when action in [:publish, :reject, :restore] do
-    valid_action?(action, implementation) and
+    valid_action?(action, implementation)  and
       Permissions.authorized?(claims, :publish_implementation, domain_id)
   end
 
@@ -209,15 +214,20 @@ defmodule TdDq.Implementations.Policy do
       |> Changeset.apply_changes()
       |> permissions()
 
-    case Changeset.fetch_change(changeset, :segments) do
+    perms = case Changeset.fetch_change(changeset, :segments) do
       :error -> perms
       {:ok, _} -> [:manage_segments | perms]
+    end
+    case Changeset.fetch_change(changeset, :status) do
+      :error -> perms
+      {:ok, :published} -> [:publish_implementation | perms]
+      {:ok, _} -> perms
     end
   end
 
   defp permissions(%Implementation{} = impl) do
     impl
-    |> Map.take([:rule_id, :segments, :implementation_type])
+    |> Map.take([:rule_id, :segments, :implementation_type, :status])
     |> Enum.flat_map(fn
       {:implementation_type, "raw"} -> [:manage_raw_quality_rule_implementations]
       {:implementation_type, "basic"} -> [:manage_basic_implementations]