+++ CHANGELOG.md
@@ -23,7 +23,7 @@
   their original `name` regardless of whether they have an alias or not. `alias`
   is indexed as an independent property if present.
 - [TD-5199] Soft delete for rule
-- [TD-5433] ignore ReferenceData enrich in Implementation if it doesn't exists
+- [TD-5433] ignore ReferenceData enrich in Implementation if it doesn't exist
 - [TD-5432] Publish rule_result_created event using implementation_ref as
   resource_id instead of implementation_id.